+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "arrowParens": "avoid",
+  "semi": true,
+  "singleQuote": true,
+  "bracketSpacing": true
+}
