--- js/head.js
@@ -1,3 +1,0 @@
-module.exports.head = head = () => {
-  window.observer = lozad(); // Instantiate, see footer for run
-};
