+++ theme/templates/base.html
@@ -9,16 +9,14 @@
   {% endblock %}
 
   <div class="footer">
-      <ul>
-        <li><a href="{{ TWITTER_URL }}">Twitter</a></li>
-        <li><a href="{{ GITHUB_URL }}">GitHub</a></li>
-        <li><a href="/resume">Resume</a></li>
-        <li><a href="/about">About</a></li>
-        <li class="last"><a href="/rss.xml">RSS</a></li>
-      </ul>
-    </div>
-
-  <script type="text/babel">twemoji.parse(document.body);</script>
+    <ul>
+      <li><a href="{{ TWITTER_URL }}">Twitter</a></li>
+      <li><a href="{{ GITHUB_URL }}">GitHub</a></li>
+      <li><a href="/resume">Resume</a></li>
+      <li><a href="/about">About</a></li>
+      <li class="last"><a href="/rss.xml">RSS</a></li>
+    </ul>
+  </div>
 </body>
