+++ theme/templates/end.html
@@ -1,8 +1,1 @@
-  <!-- // The ones we can do at the end -->
-  <script async src="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.2/highlight.min.js"></script>
-  <script async src="//www.googletagmanager.com/gtag/js?id=UA-155297229-1"></script>
-
-  <script src="/end.js"></script>
-  <script>blog.end();</script>
-
 </html>
