+++ theme/templates/resume.html
@@ -201,8 +201,6 @@
       </div>
     </div>
   </div>
-
-  <script type="text/babel">twemoji.parse(document.body);</script>
 </body>
